removed old resolveWith method
replaced by external type inference function within the dsl package
