+++ src/test/scala/dsl.scala
@@ -1,7 +1,7 @@
 package slender
 
-import dsl._
-
+import slender.dsl.implicits._
+import slender.dsl.inference._
 import org.scalatest.FunSuite
 
 object collections {
@@ -29,37 +29,37 @@ class DslTests extends FunSuite {
 
   test("Simple for-comprehension") {
     val query = For ("x" <-- stringCounts1) Collect 1
-    assert(query.isResolved)
+    assert(query.isTyped)
     assert(query.ringType == IntType)
     assert(query ==
-      Sum(stringCounts1 * {"x" ==> 1}).resolve
+      inferTypes(Sum(stringCounts1 * {"x" ==> 1}))
     )
   }
 
   test("Simple yield") {
     val query = For ("x" <-- stringCounts1) Yield "x"
-    assert(query.isResolved)
+    assert(query.isTyped)
     assert(query.ringType == MappingType(StringKeyType,IntType))
     assert(query ==
-      Sum(stringCounts1 * {"x" ==> sng("x")}).resolve
+      inferTypes(Sum(stringCounts1 * {"x" ==> sng("x")}))
     )
   }
 
   test("Predicated yield") {
     val query = For ("x" <-- intCounts iff "x" === 1) Yield "x"
-    assert(query.isResolved)
+    assert(query.isTyped)
     assert(query.ringType == MappingType(IntKeyType,IntType))
     assert(query ==
-      Sum(intCounts * {"x" ==> sng("x","x"===1)}).resolve
+      inferTypes(Sum(intCounts * {"x" ==> sng("x","x"===1)}))
     )
   }
 
   test("Flatten") {
     val query = For ("x" <-- bagOfBags) Collect fromK("x")
-    assert(query.isResolved)
+    assert(query.isTyped)
     assert(query.ringType == BagType(StringKeyType))
     assert(query ==
-      Sum(bagOfBags * {"x" ==> fromK("x")}).resolve
+      inferTypes(Sum(bagOfBags * {"x" ==> fromK("x")}))
     )
   }
 
@@ -68,7 +68,7 @@ class DslTests extends FunSuite {
       For ("k" <-- bagOfPairs) Yield (
         "k"._1 --> sng("k"._2)
       )
-    assert(query.isResolved)
+    assert(query.isTyped)
     assert(query.ringType == MappingType(StringKeyType,BagType(IntKeyType)))
   }
 
@@ -80,7 +80,7 @@ class DslTests extends FunSuite {
         ("k"._1, toK(group("k"._1)))
       )
 
-    assert(query.isResolved)
+    assert(query.isTyped)
     assert(query.ringType ==
       BagType(
         (StringKeyType, BoxedRingType(BagType(IntKeyType)))
@@ -94,9 +94,9 @@ class DslTests extends FunSuite {
         "y" --> sng("z")
       )
     )
-    assert(!query.isResolved)
-    val resolvedQuery = query.resolveWith(Map("z" -> StringKeyType))
-    assert(resolvedQuery.isResolved)
+    assert(!query.isTyped)
+    val resolvedQuery = inferTypesR(query, Map("z" -> StringKeyType))
+    assert(resolvedQuery.isTyped)
     println(resolvedQuery.ringType)
   }
